Funcionalidades funcionando.

+++ pets-data-base/src/app/tarefas/dono/dono.component.ts
@@ -1,7 +1,10 @@
 import { Component, OnInit } from '@angular/core';
+import { PetService } from '../shared/pet.service';
 import { DonoService } from '../shared/dono.service';
 import { Dono } from '../shared/dono.model';
+import { Pet } from '../shared/pet.model';
 import { NgForm } from '@angular/forms';
+import { element } from 'protractor';
 
 @Component({
   selector: 'app-dono',
@@ -12,11 +15,14 @@ import { NgForm } from '@angular/forms';
 export class DonoComponent implements OnInit {
   dono = {} as Dono;
   donos: Dono[];
+  pet = {} as Pet;
+  pets: Pet[];
 
-  constructor(private donoService: DonoService) {}
+  constructor(private donoService: DonoService, private petService: PetService) {}
   
   ngOnInit() {
     this.getDonos();
+    this.getPets();
   }
 
   /* Funções do CRUD do DONO */
@@ -41,21 +47,44 @@ export class DonoComponent implements OnInit {
     });
   }
 
+  // Chama o serviço para obtér todos os pets
+  getPets() {
+    this.petService.getPets().subscribe((pets: Pet[]) => {
+      this.pets = pets;
+    });
+  }
+
   // deleta um dono
   deleteDono(dono: Dono) {
-    this.donoService.deleteDono(dono).subscribe(() => {
-      this.getDonos();
-    });
+    if (confirm('Tem certeza que deseja deletar os dados?')) {
+      this.pets.forEach(element => {
+        if(element.id_dono == dono.id) {
+          this.petService.deletePet(element).subscribe(() => {
+
+          });
+        }
+      })
+      this.donoService.deleteDono(dono).subscribe(() => {
+        this.getDonos();
+      });
+    } else {
+      console.log('Nada foi alterado');
+    }
   }
 
   // copia o dono para ser editado.
   editDono(dono: Dono) {
-    this.dono = { ...dono };
+    if (confirm('Tem certeza que quer editar os dados?')) {
+      this.dono = { ...dono };
+    } else {
+      console.log('Nada foi alterado');
+    }
   }
 
   // limpa o formulario
   cleanForm(form: NgForm) {
     this.getDonos();
+    this.getPets();
     form.resetForm();
     this.dono = {} as Dono;
   }
