+++ pets-data-base/src/app/tarefas/pet/pet.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { PetService } from '../shared/pet.service';
+import { DonoService } from '../shared/dono.service';
 import { Pet } from '../shared/pet.model';
+import { Dono } from '../shared/dono.model';
 import { NgForm } from '@angular/forms';
 
 @Component({
@@ -8,14 +10,42 @@ import { NgForm } from '@angular/forms';
   templateUrl: './pet.component.html',
   styleUrls: ['./pet.component.css']
 })
+
 export class PetComponent implements OnInit {
   pet = {} as Pet;
   pets: Pet[];
+  dono = {} as Dono;
+  donos: Dono[];
 
-  constructor(private petService: PetService) {}
+  especies: any[] = [
+    {name:'Cachorro'},
+    {name:'Gato'}
+  ];
+
+  racasdog: any[] = [
+    {name:'Nao definida'},
+    {name:'Bulldog'},
+    {name:'Pug'},
+    {name:'Poodle'},
+    {name:'Golden'}
+  ];
+
+  racascat: any[] = [
+    {name:'Nao definida'},
+    {name:'Persa'},
+    {name:'Siames'},
+    {name:'Himalaia'},
+    {name:'Maine'}
+  ];
+
+  constructor(private petService: PetService, private donoService: DonoService) {}
 
   ngOnInit(): void {
+    
     this.getPets();
+    this.getDonos();
+    this.pet.especie = 'Cachorro';
+    this.pet.raca = 'Nao definida';
   }
 
   /* Funções do CRUD do PET */
@@ -40,21 +70,39 @@ export class PetComponent implements OnInit {
     });
   }
 
+  // Chama o serviço para obtér todos os donos
+  getDonos() {
+    this.donoService.getDonos().subscribe((donos: Dono[]) => {
+      this.donos = donos;
+      this.pet.id_dono = this.donos[0].id;
+    });
+  }
+
   // deleta um pet
   deletePet(pet: Pet) {
-    this.petService.deletePet(pet).subscribe(() => {
-      this.getPets();
-    });
+    if (confirm('Tem certeza que deseja deletar os dados?')) {
+      this.petService.deletePet(pet).subscribe(() => {
+        this.getPets();
+        this.getDonos();
+      });
+    } else {
+      console.log('Nada foi alterado');
+    }
   }
 
   // copia o pet para ser editado.
   editPet(pet: Pet) {
-    this.pet = { ...pet };
+    if (confirm('Tem certeza que deseja editar os dados?')) {
+      this.pet = { ...pet };
+    } else {
+      console.log('Nada foi alterado');
+    }
   }
 
   // limpa o formulario
   cleanForm(form: NgForm) {
     this.getPets();
+    this.getDonos();
     form.resetForm();
     this.pet = {} as Pet;
   }
